Fix libcxx-19 char_traits<uint8_t> build failure
While char_traits<uint8_t> is not officially supported by any C++
standard, gcc and libcxx(<19) have extensions to support the type. The
C++20 standard introduces support for char_traits<char8_t>.

Starting with libcxx-19, the extensions to support char_traits<T> where
T is not a type specified by a C++ standard has been dropped. See
https://reviews.llvm.org/D138307 for details.

This causes zxing-cpp to fail to build if the C++ compiler adheres to
the C++17 (or older) and uses libcxx-19 as the C++ runtime library.

The fix is to use char_traits<char8_t> when possible, fallback to
char_traits<uint8_t> when not possible, and emit a configure error if
neither type is available.

Moreover, this patch changes the preferred standard from C++17 to C++20
as C++20 defines char_traits<char8_t>.

Fixes: #822

Signed-off-by: Nicholas Vinson <nvinson234@gmail.com>