PDF417: prevent wrong position info with right side collapsing to (0,0)
Truncated PDF417 symbols don't have a stop-guard. The top-right and
bottom-right info for those was always (0,0). This could also happen for
normal PDF417 symbols when detecting the stop-guard failed.

This patch introduces a very crude approximation of the right side of the
symbol for this case that does at least make some sense.

I consider this a fix for #866. This is related to #145 and #340.