DataBar: fix inconsistency / plain spec violation (missing "(01)" prefix)
The GS1 DataBar ISO specification clearly states in Section 9 that our two
BarcodeFormats DataBar and DataBarLimited always implicitly encode the AI
"01" prefix. This has been implemented in the new DataBarLimited version
but not in the old DataBar variant.

Also to make the HRI of all DataBar variants (including the Expanded
variant) consistent, all of them are now reported to contain GS1 data and
hence be rendered as "(01)0123...." instead of "010123....".