WriteBarcode: use new BARCODE_RAW_TEXT feature from libzint
This is a significant step forward towards a complete libzint based
writer backend. It removes the use of the reader API to generate the
Barcode object after creating the symbol with libzint.

This adds a string based `options` member to `CreatorOptions` to pass
symbology specific flags and options. This is an experiment and is
motivated by the discussion in #862.

This changeset draws heavily from the work done by @gitlost in his
diagnostics2 branch.