ReedSolomonDecoder: add an extra check to validate message after
This reduces the false positives that can happen if more than the
recoverable number of errors occur by half (at least in the minimal testing
setup).

This has been discussed in #940 as a discovery separate from the original
issue. So instead of reporting a valid DataMatrix symbol with the wrong
content, zxing-cpp now reports a checksum error.