QRDetector: tune distance and angle checks for FinderPatternSets
Turned out the scaling of the squaredDistance with the square of the
finder pattern size ratio was skewing the cosine calculation too much,
which could lead to accepting degenerate triangles like a,b,c on a line.
This could cause the alignment pattern detection to find the same points
for different alignment patterns and that in turn could result in
RegressionLine creations with two identical points, which caused an
assert to fail (see second issue in #951).