fix: ZXingReaderTest in example folder may fail
The ZXingReaderTest sometimes fails, because it implicitly depends on
the ZXingWriterTest, as it needs the test.png created from this test.

This fix tells cmake explicitly that the ZXingReaderTest depends on
the ZXingWriterTest. This ensures that these unit tests are being
executed sequentially even when using ctest with the -j flag.